projects/mario_kart_steering/requirements.txt
pyautogui==0.9.41
pyserial==3.4